requirements.txt
Django==2.2.28  # Última versión LTS compatible con djongo
djangorestframework==3.11.2  # Compatible con Django 2.2
djangorestframework-simplejwt==4.7.2  # Compatible con Python 3.10 y Django 2.2
pymongo==3.12.3  # Última versión recomendada compatible con djongo
djongo==1.3.6  # Compatible con Django 2.2
dnspython==2.2.1  # Compatible con pymongo y djongo
django-cors-headers==3.2.1  # Compatible con Django 2.2
bcrypt==3.1.7  # Asegura compatibilidad con Django 2.2
django-environ==0.4.5  # Compatible con Django 2.2
pytz==2023.3  # Versión estable y requerida por Django
django-extensions==3.0.9
drf-yasg==1.21.5  # Compatible con Django 2.2 y Django REST Framework
